New Development: No             JIRA Issue: CAS-3014         Ready for Test: Yes      Interface Changes: VisCal.h What Interface Changed: (some private data changed from Int to Int64)          Test Programs: List test programs   Put in Release Notes: No              Module(s): synthesis            Description: Solve and apply counters now Int64s, now that 			the python binding can cope with them (r23480).
git-svn-id: https://svn.cv.nrao.edu/svn/casa/trunk/code/synthesis@23484 ac9c0da5-131f-0410-9009-a9c195f183e6

+++ MeasurementComponents/Calibrater.cc
@@ -1319,7 +1319,7 @@ Bool Calibrater::genericGatherAndSolve() {
   // Manage verbosity of partial channel averaging
   Vector<Bool> verb(vi.numberSpw(),True);
 
-  Vector<Int> nexp(vi.numberSpw(),0), natt(vi.numberSpw(),0),nsuc(vi.numberSpw(),0);
+  Vector<Int64> nexp(vi.numberSpw(),0), natt(vi.numberSpw(),0),nsuc(vi.numberSpw(),0);
 
   Int nGood(0);
   vi.originChunks();
